requirements.txt
google-adk>=1.0.0
playwright>=1.40.0
